refactor(vmux_agent): split AgentStrategy into core + CliAgentStrategy sub-trait
