+++ crates/vmux_agent/src/cli_trait.rs
@@ -1,0 +1,19 @@
+use std::collections::HashSet;
+use std::path::{Path, PathBuf};
+use std::time::SystemTime;
+
+use crate::McpServerConfig;
+use crate::strategy::AgentStrategy;
+
+pub trait CliAgentStrategy: AgentStrategy {
+    fn sessions_root(&self) -> PathBuf;
+    fn build_args(&self, mcp: &McpServerConfig, session_id: Option<&str>) -> Vec<String>;
+    fn build_env(&self, mcp: &McpServerConfig) -> Vec<(String, String)>;
+    fn discover_session(
+        &self,
+        cwd: &Path,
+        spawn_time: SystemTime,
+        claimed: &HashSet<String>,
+    ) -> Option<String>;
+    fn detect_end_time(&self, session_id: &str) -> bool;
+}
